Restructure how files are published, so we publish the "dist" directory
This flattens the list of files published, creating cleaner imports
As part of this, types have also been moved around
