+++ src/global-types.ts
